fixup! fixup! fixup! feat(ibm-vllm): add llm adapter
