+++ .gitignore
@@ -1,3 +1,10 @@
+scripts/ibm_vllm_generate_protos/dist
+scripts/ibm_vllm_generate_protos/dts
+scripts/ibm_vllm_generate_protos/types
+
+infra/bee-code-interpreter/*
+!infra/bee-code-interpreter/bee-code-interpreter.yaml
+
 ### Node template
 # Logs
 logs
