--- infra/bee-code-interpreter/k8s/.gitignore
@@ -1,3 +1,0 @@
-*
-!.gitignore
-!bee-code-interpreter.yaml
