+++ scripts/ibm_vllm_generate_protos/tsconfig.proto.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext",
+    "baseUrl": ".",
+    "declaration": true,
+    "emitDeclarationOnly": true,
+    "skipLibCheck": true,
+    "outDir": "dts"
+  },
+  "files": ["./types/generation.ts"]
+}
