--- src/adapters/ibm-vllm/types/fmaas/DecodingMethod.ts
@@ -1,26 +1,0 @@
-/**
- * Copyright 2024 IBM Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// Original file: src/adapters/ibm-vllm/proto/generation.proto
-
-export const DecodingMethod = {
-  GREEDY: "GREEDY",
-  SAMPLE: "SAMPLE",
-} as const;
-
-export type DecodingMethod = "GREEDY" | 0 | "SAMPLE" | 1;
-
-export type DecodingMethod__Output = (typeof DecodingMethod)[keyof typeof DecodingMethod];
