--- src/adapters/ibm-vllm/types/fmaas/ResponseOptions.ts
@@ -1,35 +1,0 @@
-/**
- * Copyright 2024 IBM Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// Original file: src/adapters/ibm-vllm/proto/generation.proto
-
-export interface ResponseOptions {
-  input_text?: boolean;
-  generated_tokens?: boolean;
-  input_tokens?: boolean;
-  token_logprobs?: boolean;
-  token_ranks?: boolean;
-  top_n_tokens?: number;
-}
-
-export interface ResponseOptions__Output {
-  input_text: boolean;
-  generated_tokens: boolean;
-  input_tokens: boolean;
-  token_logprobs: boolean;
-  token_ranks: boolean;
-  top_n_tokens: number;
-}
